Prevent deletion of deliverables if they have submissions (#324)
* [TEST] Add tests for deliverable deletion prevention

* [CODE] Prevent deliverable from being deleted if it has submissions

* [CODE] Disable delete deliverable button if submissions present

* [ACCEPTANCE-TEST] Adjust scenarios

+++ app/graphql/mutations/delete_deliverable.rb
@@ -20,6 +20,13 @@ module Mutations
         }
       end
 
+      if deliverable.has_submissions
+        return {
+          deliverable: nil,
+          errors: Types::UserError.from('This deliverable cannot be deleted as it has submissions.'),
+        }
+      end
+
       if deliverable.destroy
         {
           deliverable: deliverable,
